src/requirements.txt
pandas
mysql-connector-python
psycopg2-binary
python-dotenv==0.15.0
pandas
numpy
